Move StorageReader Interfaces
