--- src/StorageReader/ClaimControllerConfiguration.ts
@@ -1,4 +1,0 @@
-export interface ClaimControllerConfiguration {
-  readonly downloadRetryDelayInMinutes: number
-  readonly downloadMaxAttempts: number
-}
