--- src/StorageReader/IPFSConfiguration.ts
@@ -1,6 +1,0 @@
-/* tslint:disable:interface-name */
-
-export interface IPFSConfiguration {
-  readonly ipfsUrl: string
-  readonly downloadTimeoutInSeconds: number
-}
