--- src/StorageReader/ServiceConfiguration.ts
@@ -1,3 +1,0 @@
-export interface ServiceConfiguration {
-  readonly downloadIntervalInSeconds: number
-}
